Move SchedulePanel into ProfileInfo
