+++ src/components/ProfileInfo/SchedulePanel/components/RemoveScheduleDialog/removescheduledialog.scss
